requirements.txt
autogen-agentchat
autogen-ext[openai]
pdfplumber
pandas
python-dotenv